+++ src/main/kotlin/lotto/domain/LottoController.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoController {
