+++ src/main/kotlin/lotto/model/WinningLotto.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 class WinningLotto(val winningNumbers: Lotto, val bonusNumber: Int) {
     constructor(winningNumbers: List<Int>, bonusNumber: Int) : this(Lotto(winningNumbers), bonusNumber)
