+++ src/test/kotlin/lotto/WinningCalculatorTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.domain.WinningCalculator
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
